refactor: create caching package
